requirements.txt
slackbot==0.4.1
python-dotenv==0.6.2
nose==1.3.7
pep8==1.7.0
coverage==4.2
logging==0.4.9.6
requests==2.12.5
Pillow==4.0.0
